+++ WGLMakie/src/Serialization.js
@@ -1,5 +1,5 @@
 import * as Camera from "./Camera.js";
-import * as THREE from "https://cdn.esm.sh/v66/three@0.136/es2021/three.js";
+import * as THREE from "https://cdn.esm.sh/v66/three@0.157/es2021/three.js";
 
 // global scene cache to look them up for dynamic operations in Makie
 // e.g. insert!(scene, plot) / delete!(scene, plot)
@@ -239,10 +239,25 @@ function connect_uniforms(mesh, updater) {
     });
 }
 
+function convert_RGB_to_RGBA(rgbArray) {
+    const length = rgbArray.length;
+    const rgbaArray = new Float32Array((length / 3) * 4);
+
+    for (let i = 0, j = 0; i < length; i += 3, j += 4) {
+        rgbaArray[j] = rgbArray[i]; // R
+        rgbaArray[j + 1] = rgbArray[i + 1]; // G
+        rgbaArray[j + 2] = rgbArray[i + 2]; // B
+        rgbaArray[j + 3] = 1.0; // A
+    }
+
+    return rgbaArray;
+}
+
+
 function create_texture(data) {
     const buffer = data.data;
     if (data.size.length == 3) {
-        const tex = new THREE.DataTexture3D(
+        const tex = new THREE.Data3DTexture(
             buffer,
             data.size[0],
             data.size[1],
@@ -253,26 +268,35 @@ function create_texture(data) {
         return tex;
     } else {
         // a little optimization to not send the texture atlas over & over again
-        const tex_data =
-            buffer == "texture_atlas" ? TEXTURE_ATLAS[0].value : buffer;
+        let tex_data;
+        if (buffer == "texture_atlas") {
+            tex_data = TEXTURE_ATLAS[0].value;
+        } else {
+            tex_data = buffer;
+        }
+        let format = THREE[data.three_format];
+        if (data.three_format == "RGBFormat") {
+            tex_data = convert_RGB_to_RGBA(tex_data);
+            format = THREE.RGBAFormat;
+        }
         return new THREE.DataTexture(
             tex_data,
             data.size[0],
             data.size[1],
-            THREE[data.three_format],
+            format,
             THREE[data.three_type]
         );
     }
 }
 
 function re_create_texture(old_texture, buffer, size) {
+    let tex;
     if (size.length == 3) {
-        const tex = new THREE.DataTexture3D(buffer, size[0], size[1], size[2]);
+        tex = new THREE.Data3DTexture(buffer, size[0], size[1], size[2]);
         tex.format = old_texture.format;
         tex.type = old_texture.type;
-        return tex;
     } else {
-        return new THREE.DataTexture(
+        tex = new THREE.DataTexture(
             buffer,
             size[0],
             size[1] ? size[1] : 1,
@@ -280,6 +304,17 @@ function re_create_texture(old_texture, buffer, size) {
             old_texture.type
         );
     }
+    tex.minFilter = old_texture.minFilter
+    tex.magFilter = old_texture.magFilter
+    tex.anisotropy = old_texture.anisotropy
+    tex.wrapS = old_texture.wrapS
+    if (size.length > 1) {
+        tex.wrapT = old_texture.wrapT
+    }
+    if (size.length > 2) {
+        tex.wrapR = old_texture.wrapR
+    }
+    return tex
 }
 function BufferAttribute(buffer) {
     const jsbuff = new THREE.BufferAttribute(buffer.flat, buffer.type_length);
@@ -368,6 +403,7 @@ function create_material(program) {
         fragmentShader: program.fragment_source,
         side: is_volume ? THREE.BackSide : THREE.DoubleSide,
         transparent: true,
+        glslVersion: THREE.GLSL3,
         depthTest: !program.overdraw.value,
         depthWrite: !program.transparency.value,
     });