+++ src/layouting/transformation.jl
@@ -362,7 +362,8 @@ apply_transform(f::NTuple{3, typeof(identity)}, r::Rect) = r
 const pseudolog10 = ReversibleScale(
     x -> sign(x) * log10(abs(x) + 1),
     x -> sign(x) * (exp10(abs(x)) - 1);
-    limits=(0f0, 3f0)
+    limits=(0f0, 3f0),
+    name=:pseudolog10
 )
 
 Symlog10(hi) = Symlog10(-hi, hi)
@@ -383,7 +384,7 @@ function Symlog10(lo, hi)
     else
         x
     end
-    ReversibleScale(forward, inverse; limits=(0f0, 3f0))
+    return ReversibleScale(forward, inverse; limits=(0.0f0, 3.0f0), name=:Symlog10)
 end
 
 inverse_transform(::typeof(identity)) = identity
@@ -406,27 +407,40 @@ end
 ################################################################################
 
 """
-    Polar(theta_0::Float64 = 0.0, direction::Int = +1)
+    Polar(theta_0::Float64 = 0.0, direction::Int = +1, r0::Float64 = 0)
 
 This struct defines a general polar-to-cartesian transformation, i.e.,
 ```math
-(r, theta) -> (r \\cos(direction * (theta + theta_0)), r \\sin(direction * (theta + theta_0)))
+(r, θ) -> ((r - r₀) ⋅ \\cos(direction ⋅ (θ + θ₀)), (r - r₀) ⋅ \\sin(direction \\cdot (θ + θ₀)))
 ```
 
 where theta is assumed to be in radians.
 
-`direction` should be either -1 or +1, and `theta_0` may be any value.
+`direction` should be either -1 or +1, `r0` should be positive and `theta_0` may be any value.
+
+Note that for `r0 != 0` the inversion may return wrong results.
 """
 struct Polar
+    theta_as_x::Bool
     theta_0::Float64
     direction::Int
-    Polar(theta_0 = 0.0, direction = +1) = new(theta_0, direction)
+    r0::Float64
+    function Polar(theta_as_x = true, theta_0 = 0.0, direction = +1, r0 = 0)
+        return new(theta_as_x, theta_0, direction, r0)
+    end
 end
 
 Base.broadcastable(x::Polar) = (x,)
 
 function apply_transform(trans::Polar, point::VecTypes{2, T}) where T <: Real
-    y, x = point[1] .* sincos((point[2] + trans.theta_0) * trans.direction)
+    if trans.theta_as_x
+        r = max(0.0, point[2] - trans.r0)
+        θ = trans.direction * (point[1] + trans.theta_0)
+    else
+        r = max(0.0, point[1] - trans.r0)
+        θ = trans.direction * (point[2] + trans.theta_0)
+    end
+    y, x = r .* sincos(θ)
     return Point2{T}(x, y)
 end
 
@@ -443,11 +457,20 @@ function apply_transform(f::Polar, point::VecTypes{N2, T}) where {N2, T}
 end
 
 function inverse_transform(trans::Polar)
-    return Makie.PointTrans{2}() do point
-        typeof(point)(
-            hypot(point[1], point[2]),
-            mod(trans.direction * atan(point[2], point[1]) - trans.theta_0, 0..2pi)
-        )
+    if trans.theta_as_x
+        return Makie.PointTrans{2}() do point
+            typeof(point)(
+                mod(trans.direction * atan(point[2], point[1]) - trans.theta_0, 0..2pi),
+                hypot(point[1], point[2]) + trans.r0
+            )
+        end
+    else
+        return Makie.PointTrans{2}() do point
+            typeof(point)(
+                hypot(point[1], point[2]) + trans.r0,
+                mod(trans.direction * atan(point[2], point[1]) - trans.theta_0, 0..2pi)
+            )
+        end
     end
 end